GitHub Flat: add Deno post-processing script
the .member and .relation fields of each card have values (perhaps?)
used by the landscape app, however when importing via Cypher we end
up with Member type of "false", as well as Relation types of "member"
and "false" - as well as thousands of relationships.

This commit uses the flat action's postprocessing step to remove them
from the input data.